consolidate with existing utils class
